extract code to format taxonomy data into tree structured format
