+++ src/components/index.js
@@ -1,0 +1,3 @@
+import { structure_taxonomy_terms } from './Widgets/taxonomyUtils';
+
+export { structure_taxonomy_terms };
